requirements.txt
asgiref==3.11.1
Django==5.1.7
django-filter==24.3
django-cors-headers==4.6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
python-json-logger==2.0.7
sqlparse==0.5.5
